requirements.txt
attrs==21.4.0
click==8.1.2
flake8==4.0.1
flake8-bugbear==22.3.23
flake8-builtins==1.5.3
flake8-commas==2.1.0
flake8-polyfill==1.0.2
Flask==2.1.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mccabe==0.6.1
packaging==21.3
passlib==1.7.4
pep8-naming==0.12.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.8
pytest==7.1.1
python-dotenv==0.20.0
SQLAlchemy==1.4.35
tomli==2.0.1
Werkzeug==2.1.1
zipp==3.8.0
